+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - iOS bottom toolbar items not grouped together
+- macOS principal toolbar items not grouped together
 
 ## [1.0.0+alpha.7] - 2026-03-14